Moved inlined methods before they are called first time
Compiler only can inline method if it is implemented before it is used the first time.

Also there was cast from Pointers to Integers and visa versa, now that code Uses NativeInt.

Don't know does these changes code faster or not, if not changed methods shouldn't, most likely, be inlined.